+++ lib/main.js
@@ -35,6 +35,7 @@ class Airtable {
                 url: `${this.apiURL}/${this.baseId}/${this.tableId}${endpoint}`,
                 apiKey: this.apiKey,
                 method: method,
+                body,
             });
         });
     }
@@ -57,7 +58,7 @@ class Airtable {
             });
         });
     }
-    updateRecords(records) {
+    updateRecords(records, options) {
         return __awaiter(this, void 0, void 0, function* () {
             return this.request({ endpoint: '/', method: 'PATCH', body: { records } });
         });
